ENH Can now train DPMixtureModel with enforced sparsity (at most L of the K states can be non-zero for any data atom's assignments). Works with Gauss and ZeroMeanGauss currently. TODO: extend to Mult, Bern, and other likelihoods.

+++ bnpy/util/SparseRespUtilX.pyx
@@ -9,7 +9,7 @@ http://nealhughes.net/cython1/
 
 import numpy as np
 
-def calcSpRData_cython(double[:,:] R, int[:,:] ColIDs, int nnzPerRow):
+def calcSpRData_cython(double[:,:] R, long[:,:] ColIDs, int nnzPerRow):
     """
     """
     cdef int N = R.shape[0]
